// src/app/models/contraparte.model.ts
export class ContraparteModel {
    _id: string;
    contNumeroIdentificacion: String;
    contNombre: String;
    contOrden: String;
    contEdad: String;
    contAntiguedad: String;
    contPagare: Boolean;
    contCliente: Boolean;
    contProveedor: Boolean;
    contEmpleado: Boolean;
    contDepartamentos: String;
    contMunicipios: String;
    contEmail: String;
    contObservacion: String;
       
    }